website: basic jekyll-based static content
This creates jekyll-based website content to serve www.osbuild.org. The
layout and style is kept very simple and minimal and is open for
rewrite by people with better designing skills than me. However, this
serves as a first landing page until we get something better.

This commit imports a single landing-page `index.md`, which contains a
short introduction and links to external resources.