ci: create preview branch
This introduces a new GitHub workflow that automatically triggers on any
ref change (branches or tags). It will then run jekyll on that ref and
push the rendered preview to the `preview` branch in a subdirectory
named after the ref:

    ./preview/refs/heads/<branchname>
    ./preview/refs/tags/<tagname>

    (The jekyll baseurl is updated accordingly.)

If you now point your github-pages to the `preview` branch, you will get
a rendered preview of all branches and tags via GitHub pages.

With this enabled, we can more easily get previews of PRs and other
experimental changes before we merge it. Furthermore, we have a branch
that contains the rendered HTML and allows us to inspect it
retroactively.

As an example, this PR is sourced from `dvdhrm/osbuild.github.io`, where
I enabled gh-pages on the `preview` branch. Therefore, this PR can be
previewed at:

    https://dvdhrm.github.io/osbuild.github.io/refs/heads/cipreview

    (Note that the preview workflows take a few minutes.)