ci: add jekyll test-run
Add a new workflow that runs jekyll on all branches and PRs. This
verifies the jekyll generation works, and thus serves as neat small CI
test for us.

We cannot reproduce the exact environment that github uses to generate
the pages. However, this at least verifies the latest stable jekyll
release works.

While at it, run htmlproofer as well, to verify links and html
validity.