ci: avoid `master` for man-page imports
We import man-pages from upstream osbuild and osbuild-composer projects.
We used `master` as target branch for `osbuild`, but always intended to
switch away to it and instead use the latest release (as we already do
for composer).

Switch over to using the latest release tag instead of `master`, so the
man-page imports work again.